Create entry points for each import script
In the sw config file importScripts are defined relative to the root of
the client folder. These scripts will be requested by the service worker
using the importScripts API, so we create an entry point for each of
them. This makes each script isolated from the other, so they can be
imported in any order and run with their own bootstrap code.